refactor: main 파일 나누기
