+++ RockPaperScissors/RockPaperScissors/Display.swift
@@ -1,0 +1,13 @@
+//
+//  Display.swift
+//  RockPaperScissors
+//
+//  Created by MAC2020 on 2023/11/23.
+//
+
+import Foundation
+
+
+func displayRockScissorsPaperMenu() {
+    print(StringNameSpace.rockScissorsPaperExit.rawValue)
+}
